detect when belongs_to doesn't use primary key instead, /cc #4
We should be able to get away with this instead of dumping out the
record with a different id. When we detect that a belongs_to
association has a :primary_key declared and it isn't the reflection
class's primary key, avoid key conversion altogether and just write
the raw attribute value. The destination of the belongs_to is
guaranteed to be dumped with the corresponding value, it's just that
its id may change in the process.

This is failing under AR 2.2 because this isn't supported:

    belongs_to :something, :primary_key => 'not_id'